+++ src/bin/terra_query.rs
@@ -23,7 +23,11 @@ async fn run() -> anyhow::Result<()> {
     let terra = cli_helpers::lcd_no_tx_from_args(&cli)?;
     let contract = cli_helpers::get_arg_value(&cli, "contract")?;
     let json_str = cli_helpers::get_arg_value(&cli, "json")?;
-    let json: serde_json::Value = cli_helpers::get_json_block_expanded(json_str, None)?;
+    let secp = secp256k1::Secp256k1::new();
+    let wallet = cli_helpers::wallet_opt_from_args(&cli);
+    let seed = cli_helpers::seed_from_args(&cli);
+    let json: serde_json::Value =
+        cli_helpers::get_json_block_expanded(json_str, None, &secp, wallet, seed)?;
 
     //    let json: serde_json::Value = serde_json::from_str(json_str)?;