+++ terra-rust-cli/Changelog.md
@@ -1,5 +1,8 @@
 # Changes
 ## 1.0
+### 1.0.10 - 15-Mar-2022
+* seed_from_args / wallet_opt_from_args functions
+* expand block now expands A:key and O:key to account/operator address for key in wallet
 ### 1.0.9 - 15-Mar-2022
 * ###SENDER### will only be replaced if sender is specified. Tighten Regex
 ### 1.0.7 - 13-Mar-2022